#2047 Added backend for FloorUnions

+++ database/migrations/2023_09_28_140536_create_floor_unions_table.php
@@ -17,8 +17,8 @@ class CreateFloorUnionsTable extends Migration
             $table->id();
             $table->integer('floor_id');
             $table->integer('target_floor_id');
-            $table->integer('lat');
-            $table->integer('lng');
+            $table->float('lat');
+            $table->float('lng');
             $table->float('size');
             $table->float('rotation');
